Add list_options in Interface:classes will inherit
